Create hundred_money_and_hundred_chiekens.py
In the fifth century AD, the ancient Chinese mathematician Zhang Qiujian put forward the "Hundred Chicken Problem" in his book "Calculations": a chicken is worth five, a female chicken is worth three, and three chickens are worth one. Buy a hundred chickens for a hundred dollars, and ask how are the chickens, mothers, and chicks?