Craps_gambling_game.py
CRAPS, also known as Craps, is a very popular table gambling game in Las Vegas, USA. The game uses two dice, and the player wins points by rolling the two dice to play the game. The simple rule is: if the player rolls the dice for the first time if they roll 7 or 11 points, the player wins; if the player rolls 2 points, 3 points or 12 points for the first time, the dealer wins; other players continue to roll the dice, If the player rolls 7 points, the dealer wins; if the player rolls the first point, the player wins; for other points, the player continues to roll the dice until the winner is determined.