Create Fibonacci_sequence(first20terms).py
The Fibonacci sequence (Fibonacci sequence), also known as the golden section sequence, is the Italian mathematician Leonardo Fibonacci (Leonardoda Fibonacci) proposed a rabbit under ideal assumptions in the "Book of Calculations" The number sequence introduced by the problem of growth rate, so this number sequence is also dubbed the "rabbit sequence". The characteristic of the Fibonacci sequence is that the first two numbers of the sequence are both 1, and starting from the third number, each number is the sum of the previous two numbers, like: 1, 1, 2, 3, 5 , 8, 13, 21, 34, 55, 89, 144, .... The Fibonacci sequence has direct applications in modern physics, quasi-crystal structure, chemistry and other fields.