Create perfect_number.py
A perfect number is also called a perfect number or a complete number. The sum of all its true factors (that is, factors other than itself) (that is, the factor function) is exactly equal to itself. For example: 6 ($6=1+2+3$) and 28 ($28=1+2+4+7+14$) are perfect numbers. Perfect numbers have many magical characteristics, and those who are interested can learn about them on their own.