Create Function_parameters.py
A function is a "building block" of a code supported by most programming languages, but there are still many differences between functions in Python and functions in other languages. One of the notable differences is Python's function parameters deal with. In Python, function parameters can have default values ​​and support the use of variable parameters, so Python does not need to support function overloading like other languages, because we can define a function with many different How to use, the following are two small examples.